Display anyhow error chains better
The default Display / `{}` formatter only shows the outermost error. We
have to use `{:#}` to show all the errors in the chain. This will make
stream errors somewhat more informational.

Before:

    Stream error: Storage error: service error

After:

    Stream error e=Storage error: service error: NoSuchKey: The specified key does not exist.: NoSuchKey: The specified key does not exist.

(after I manually mucked with the DB to change the S3 url for an item to
a non-existent name)